+ order-service refactor + Intergration test across all micro-services

+++ order-service/src/main/java/com/sellersphere/orderservice/data/OrderItem.java
@@ -1,4 +1,4 @@
 package com.sellersphere.orderservice.data;
 
-public record OrderItem(String productId, String productName, int quantity, double price) {
+public record OrderItem(String productId, String productName, int quantity, int price) {
 }
